refactor(types): observability DTO 上移到 src/types/ 解开内部循环
把 observability/ 下原本散落在多个文件的纯类型声明集中到 src/types/observability.ts，源文件改为 import type + re-export 的 facade 模式：

- skill-chain.ts / skill-chain-advisories.ts / review-state.ts / problem-patterns.ts / inbox.ts / experience.ts / experience-frontmatter.ts / trace-source.ts 全部改为从 types/ 拉类型
- experience.ts 与 inbox.ts、experience-frontmatter.ts 之间的 type-only 循环依赖通过提取到 leaf 层解开
- 运行时函数、常量、内部类型 (ExperienceEpisodeRange、aggregateExperienceChecklistItemStatus) 保留在原位
- 对外导出名（包括所有 Experience\*、Observation\* DTO）与重构前一致，无 BREAKING-COMPARABILITY
- types/observability.ts 内 ObservationSkillChain 复用 shared/hard-rules.ts 的 SkillHardRule / SkillWorkflow，避免重复定义

测量学红区文件 (report.ts / bootstrap.ts / statistics.ts / comparability.ts / judge.ts / assertions.ts / layered-scores.ts) 零改动；judge-hash-frozen / llm-enhanced-review-prompt / html-renderer snapshot 全绿；1615 测试通过。
