+++ src/types/index.ts
@@ -6,3 +6,4 @@ export * from './report.js';
 export * from './storage.js';
 export * from './doctor.js';
 export * from './diagnosis.js';
+export * from './observability.js';
